move database layer into shared
